requirements.txt
pyarrow==14.0.1
mlflow==2.9.2
scikit-surprise==1.1.3
psycopg2
neptune
